Move functions to cc file
